// src/types/partials/OperationDefinitionParameter.ts
/* eslint-disable no-unused-vars */
/* eslint-disable no-trailing-spaces */
// This file is auto-generated by generate_types so do not edit manually

import { TExtension } from '../partials/Extension';
import { TInt } from '../simpleTypes/Int';
import { TCanonical } from '../simpleTypes/Canonical';
import { TOperationDefinitionBinding } from '../partials/OperationDefinitionBinding';
import { TOperationDefinitionReferencedFrom } from '../partials/OperationDefinitionReferencedFrom';

export type TOperationDefinitionParameter = {
    id?: String;
    extension?: TExtension[];
    modifierExtension?: TExtension[];
    name: String;
    use: String;
    min: TInt;
    max: String;
    documentation?: String;
    type?: String;
    targetProfile?: TCanonical[];
    searchType?: String;
    binding?: TOperationDefinitionBinding;
    referencedFrom?: TOperationDefinitionReferencedFrom[];
    part?: TOperationDefinitionParameter[];
};